[bugfix] ChecksumMismatch for /etc/sudoers.d/rails
problem:  before it was not possible to have multiple commands being added to the sudoers file of the rails user.  only one command would win which would end up in the sudoers file changing and mismatches appearing that caused chef-client runs to fail

solution:  we add all sudo commands for the deploy user to one variable and write the sudoers file at the end.